Remove libc::signal calls from FdPtyPair::prepare_process
On execve(2), handled signals are reset to SIG_DFL. Signals which are
set to SIG_IGN will not be reset. Thus, we shouldn't touch any signals
we didn't explicitly set ourselves. SIGCHLD may be an exception to this
rule, but for the time being since we don't do anything with it, let's
not touch it either.